+++ lib/super_validation_a.dart
@@ -10,33 +10,73 @@ abstract class SuperValidationA {
   bool get isValid => validation == null;
 }
 
-abstract class SuperValidationValue<T> extends SuperValidationA {
-  Stream<T?> get streamValue;
-  T? get value;
-  set value(T? value);
+typedef SuperValidationValueFunc<T> = String? Function(T? value);
 
+class SuperValidationValue<T> extends SuperValidationA {
+  SuperValidationValueFunc<T>? validateFunc;
   final SuperValidationStore<T>? store;
-  SuperValidationValue({this.store}) {
+  SuperValidationValue({
+    this.validateFunc,
+    this.store,
+  }) {
     if (store != null) {
-      value = store!.valueStored;
-    }
-    _streamSubscription = streamValue.listen((T? value) {
-      if (store != null) {
-        store!.valueStored = value;
+      _streamSubscription = streamValue.listen((value) {
+        store?.valueStored = value;
+      });
+      final stored = store?.valueStored;
+      if (stored != null) {
+        value = stored;
       }
-    });
+    }
   }
+
   StreamSubscription<T?>? _streamSubscription;
 
   @mustCallSuper
   Future<void> dispose() async {
     await _streamSubscription?.cancel();
+    await validationController.close();
+    await valueController.close();
   }
 
-  void save() {
-    if (store != null) {
-      store!.valueStored = value;
+  @override
+  String? get validation => internalValidation;
+  @override
+  set validation(String? value) {
+    if (internalValidation == value) {
+      return;
     }
+    validationController.add(value);
+    internalValidation = value;
+  }
+
+  @protected
+  String? internalValidation;
+  @protected
+  final StreamController<String?> validationController =
+      StreamController<String?>.broadcast();
+  @override
+  Stream<bool> get streamIsValid =>
+      validationController.stream.map((event) => event == null);
+
+  @override
+  Stream<String?> get streamValidation => validationController.stream;
+
+  Stream<T?> get streamValue => valueController.stream;
+  @protected
+  final StreamController<T?> valueController = StreamController<T?>.broadcast();
+  @protected
+  T? internalValue;
+
+  T? get value => internalValue;
+
+  set value(T? value) {
+    if (internalValue == value) {
+      return;
+    }
+    internalValue = value;
+    valueController.add(value);
+    validation = validateFunc?.call(value);
   }
 }